WIP: Moved Packing to its own file
